fix: multiplayer

+++ src/Common/DTO/gameCommand.cpp
@@ -1,6 +1,6 @@
 #include "gameCommand.h"
 
-GameCommandDTO::GameCommandDTO(const uint32_t& player_id, 
+GameCommandDTO::GameCommandDTO(const uint8_t& player_id,
     const Direction& moveType, const Command& cmd):
         CommandDTO(player_id, cmd), moveType(moveType) {}
