+++ src/Common/DTO/gameCommand.h
@@ -12,7 +12,7 @@ private:
     Direction moveType;
 
 public:
-    GameCommandDTO(const uint32_t& player_id, const Direction& moveType, const Command& cmd);
+    GameCommandDTO(const uint8_t& player_id, const Direction& moveType, const Command& cmd);
 
     Direction getMoveType() const;
